Move the limited-threads test to its own file
