+++ test/tbb-preview/resource_limiting_test.cpp
@@ -2,13 +2,11 @@
 #include "phlex/utilities/thread_counter.hpp"
 
 #include "catch2/catch_test_macros.hpp"
-#include "fmt/std.h"
 #include "oneapi/tbb/flow_graph.h"
 #include "spdlog/sinks/basic_file_sink.h"
 #include "spdlog/spdlog.h"
 
 #include <atomic>
-#include <semaphore>
 #include <string>
 #include <string_view>
 #include <tuple>
@@ -46,89 +44,7 @@ namespace {
     spdlog::set_default_logger(logger);
   }
 
-  template <typename F, typename... Args>
-  class managed_thread {
-    enum class states : std::uint8_t { off, drive, park };
-
-    F f_;
-    std::optional<std::tuple<Args...>> args_;
-    std::jthread thread_;
-    std::binary_semaphore work_ready_{0};
-    std::binary_semaphore slot_ready_{0};
-    std::atomic<states> gear_ = states::off;
-
-    void run(std::stop_token st)
-    {
-      while (true) {
-        work_ready_.acquire();
-        if (gear_ == states::park || st.stop_requested()) {
-          break;
-        }
-
-        std::apply(f_, args_.value());
-        slot_ready_.release();
-      }
-    }
-
-  public:
-    managed_thread(F f) : f_{std::move(f)}, thread_{[this](std::stop_token st) { run(st); }} {}
-
-    ~managed_thread()
-    {
-      gear_ = states::park;
-      work_ready_.release();
-    }
-
-    void execute(Args const&... args)
-    {
-      if (gear_ == states::off) {
-        gear_ = states::drive;
-      }
-
-      args_ = std::make_tuple(args...);
-      work_ready_.release();
-      slot_ready_.acquire();
-      args_.reset();
-    }
-  };
-
 } // namespace
-
-TEST_CASE("std::threads as limited resources")
-{
-  auto f = [](unsigned int const i) {
-    spdlog::info("Executing on thread {} with argument {}", std::this_thread::get_id(), i);
-  };
-  using managed_thread_type = managed_thread<decltype(f), unsigned int>;
-
-  managed_thread_type t1{f};
-  managed_thread_type t2{f};
-  flow::resource_limiter<managed_thread_type*> pinned_thread_resource{&t1, &t2};
-
-  flow::graph g;
-  unsigned int i{};
-  flow::input_node src{g, [&i](flow_control& fc) {
-                         if (i < 10) {
-                           return ++i;
-                         }
-                         fc.stop();
-                         return 0u;
-                       }};
-
-  flow::resource_limited_node<unsigned int, std::tuple<>> geant4_node{
-    g,
-    flow::unlimited,
-    std::tie(pinned_thread_resource),
-    [](unsigned int const i, auto&, managed_thread_type* geant4_resource) {
-      spdlog::info("Launching from thread {} with argument {}", std::this_thread::get_id(), i);
-      geant4_resource->execute(i);
-    }};
-
-  make_edge(src, geant4_node);
-
-  src.activate();
-  g.wait_for_all();
-}
 
 TEST_CASE("Serialize functions based on resource", "[multithreading]")
 {
